Move newsletter dashboard form to newsletters app
